Move controller actions to service, add console command, only verify key if not console request
