+++ src/DbDump.php
@@ -12,6 +12,7 @@ namespace swishdigital\dbdump;
 
 use swishdigital\dbdump\variables\DbDumpVariable;
 use swishdigital\dbdump\models\Settings;
+use swishdigital\dbdump\services\DbDumpService;
 
 use Craft;
 use craft\base\Plugin;
@@ -57,8 +58,18 @@ class DbDump extends Plugin
      */
     public function init()
     {
+		if (Craft::$app->getRequest()->getIsConsoleRequest()) {
+            $this->controllerNamespace = 'db-dump\\console\\default';
+        } else {
+			$this->controllerNamespace = 'db-dump\\default';
+		}
+
         parent::init();
         self::$plugin = $this;
+
+		$this->setComponents([
+			'DbDumpService' => DbDumpService::class,
+		]);
 
         Event::on(
             UrlManager::class,
